Upgrades
Updates:
all angular from ~4.0.0 to ~4.1.0
both nativescript-angular  and tns-core-modules from ~1.5.1 to ~3.0.0
added lazy 1.0.11
nativescript-dev-sass from ^1.1.0 to ^1.1.1
nativescript-dev-typescript from ~0.3.5 to ~0.4.5
nativescript-dev-webpack from ~0.4.0 to ~0.5.0
typescript from ~2.1.0 to ~2.2.1
webpack from ~2.3.3 to ~2.5.1

nativescript-dev-webpack now uses
webpack.config.js
so, I've removed the three previous files

I've made some changes / added to App_Resources/Android:
** Keep in mind that the string nativescript.id in package.json must be
equal to the string manifest.package in AndroidManifest.xml
In case not, the build will compile correctly, but the app will crash on
the deivce. (Maybe handy to add this to the README.md?)
.... I've added the /App_Resources/Android/values/ strings.xml file:
The name of the application on the device.